Add new UI for the calculator interactions. Implementation of basic functionality for calculator. Hitting number buttons adds numbers, hitting operation keys moves number to history and allows user to press more numbers, hitting equals solves the problems.
Note: This does not allow the user to do multiple operations in a row or
      type to enter numbers.

QCalcWidget[.h, .cpp]
New Class implementing the logic behind the calculator buttons. Includes
connections to the ButtonPressed Signal.

QCalcWidget.ui
New UI Form containing the calculator layout.

calculator.pro
Added new files

main.cpp
Whitespace changes with the update to using clang-format on save.

mainwindow.ui
Update the main window to contain a QCalcWidget.
This main window is what is shown when the application is launched, so now
the calculator is is what is shown when the application is launched.

.clang-format
New clang format file used for all cpp/h files.